Separated api and api-auth
api-auth is now a package containing only the NextAuth config. We need to do this because it can't use type:module and thus break the frontend build. But the bot requires API and requires the type module.
Fixed the import on frontend to only import from `dist` because there is an issue importing from @echo/api directly atm.